+++ src/components/VideoStepControl/VideoStepControl.tsx
@@ -1,3 +1,4 @@
+import { useState, useLayoutEffect } from "react";
 import { IconButton } from "@chakra-ui/react";
 
 import { PlayerTrackPrev as PlayerTrackPrevIcon, PlayerTrackNext as PlayerTrackNextIcon } from "tabler-icons-react";
@@ -9,24 +10,33 @@ interface Props {
 }
 
 export default function VideoStepControl({ onClick, frameRate, direction }: Props) {
+  const [mouseDown, setMouseDown] = useState(false);
   const [value, Icon] = direction === "forwards" ? [1, <PlayerTrackNextIcon />] : [-1, <PlayerTrackPrevIcon />];
   const frameLength = 1 / frameRate;
 
-  function handleClick(event: React.MouseEvent<HTMLButtonElement>) {
-    const distance = (() => {
-      if (event.getModifierState("Control")) {
-        return 1;
+  useLayoutEffect(() => {
+    if (mouseDown === true) {
+      onClick(frameLength * value);
+    }
+
+    const interval = setInterval(() => {
+      if (mouseDown === true) {
+        onClick(frameLength * value);
       }
+    }, 100);
 
-      if (event.getModifierState("Shift")) {
-        return 10;
-      }
+    return () => {
+      clearInterval(interval);
+    };
+  }, [mouseDown]);
 
-      return frameLength;
-    })();
-
-    onClick(distance * value);
+  function handleMouseDown() {
+    setMouseDown(true);
   }
 
-  return <IconButton icon={Icon} aria-label="Step backwards" onClick={handleClick} />;
+  function handleMouseUp() {
+    setMouseDown(false);
+  }
+
+  return <IconButton icon={Icon} aria-label="Step backwards" onMouseDown={handleMouseDown} onMouseUp={handleMouseUp} />;
 }