+++ src/pages/App.tsx
@@ -50,7 +50,7 @@ export default function App() {
               VODON PRO
             </Heading>
             <Flex as={"nav"} ml={"4"}>
-              <NavLink to="/">Setup videos</NavLink>
+              <NavLink to="/">Setup</NavLink>
               <NavLink to="/review">Review</NavLink>
               <NavLink to="/settings">Settings</NavLink>
               <NavLink to="/about">About</NavLink>