+++ src/pages/SetupVideos.tsx
@@ -3,22 +3,31 @@ import { css } from "@emotion/react";
 
 import {
   Box,
+  Button,
+  ButtonGroup,
   Flex,
   Grid,
   GridItem,
-  Select,
   Heading,
+  Kbd,
   ListItem,
+  Modal,
+  ModalBody,
+  ModalCloseButton,
+  ModalContent,
+  ModalFooter,
+  ModalHeader,
+  ModalOverlay,
   OrderedList,
   UnorderedList,
-  Kbd,
 } from "@chakra-ui/react";
 
 import useStore from "../services/store";
 
+import { Help as HelpIcon, X as XIcon } from "tabler-icons-react";
+
 import VideoAdd from "../components/VideoAdd/VideoAdd";
 import VideoAligner from "../components/VideoAligner/VideoAligner";
-import WithSidebar from "../layouts/WithSidebar";
 
 const addVideoCellStyles = css`
   aspect-ratio: 16 / 9;
@@ -36,8 +45,11 @@ const addVideoCellStyles = css`
 `;
 
 export default function SetupVideos() {
+  const showSetupInstructions = useStore((state) => state.showSetupInstructions);
   const videos = useStore((state) => state.videos);
-  const maxDuration = useStore((state) => state.maxDuration);
+
+  const setShowSetupInstructions = useStore((state) => state.setShowSetupInstructions);
+  const clearVideos = useStore((state) => state.clearVideos);
 
   const [rowCount, setRowCount] = useState("2");
 
@@ -49,56 +61,41 @@ export default function SetupVideos() {
     );
   });
 
-  // TODO: Think about bringing this back if needed
-  function handleRowCountChange(event: React.ChangeEvent<HTMLSelectElement>) {
-    setRowCount(event.target.value);
+  function handleCloseInstructions() {
+    setShowSetupInstructions(false);
   }
 
-  const renderedSidebar = (
-    <Box p={4} color={"whiteAlpha.600"}>
-      <Heading color={"white"} fontSize={"lg"} mb={"4"}>
-        Getting started
-      </Heading>
-      <OrderedList my={"4"}>
-        <ListItem mb={"4"}>Either click to add a video or drag and drop them from your desktop.</ListItem>
-        <ListItem mb={"4"}>
-          Pick a point in time that is easy to identify in all the videos (countdowns before a round start works well
-          here)
-        </ListItem>
-        <ListItem mb={"4"}>Align all the videos to the same point in time using the controls on each video.</ListItem>
-        <ListItem mb={"4"}>When all videos are aligned, go to the "Review" tab to start the VOD review</ListItem>
-      </OrderedList>
+  function handleShowInstructions() {
+    setShowSetupInstructions(true);
+  }
 
-      <Heading color={"white"} fontSize={"lg"} mb={"4"} mt={"8"}>
-        Hints
-      </Heading>
-      <UnorderedList my={"4"}>
-        <ListItem mb={"4"}>The arrow controls move you back and forward by one frame</ListItem>
-        <ListItem mb={"4"}>
-          Hold <Kbd>Control</Kbd> when clicking the arrow controls to move forward and back by one second
-        </ListItem>
-        <ListItem mb={"4"}>
-          Hold <Kbd>Shift</Kbd> when clicking the arrow controls to move forward and back by ten seconds
-        </ListItem>
-      </UnorderedList>
-
-      {/*
-      <Heading fontSize={'lg'} my={'4'}>Controls</Heading>
-      <Flex align={"center"}>
-        <label style={{ whiteSpace: "nowrap" }}>Videos per row:</label>
-        <Select value={rowCount} onChange={handleRowCountChange} ml={4}>
-          <option value="1">1</option>
-          <option value="2">2</option>
-          <option value="3">3</option>
-        </Select>
-      </Flex>
-      */}
-    </Box>
-  );
+  function handleClearVideos() {
+    if (confirm("Are you sure you want to remove all videos?")) {
+      clearVideos();
+    }
+  }
 
   return (
-    <WithSidebar sidebar={renderedSidebar}>
+    <>
       <Box overflowY={"auto"} height={"calc(100vh - 5rem)"} width={"100%"}>
+        <Flex
+          boxSizing={"border-box"}
+          borderBottom={"1px"}
+          borderColor={"whiteAlpha.300"}
+          height={"4rem"}
+          px={8}
+          alignItems={"center"}
+          justifyContent={"flex-start"}
+        >
+          <ButtonGroup size={"sm"}>
+            <Button colorScheme={"cyan"} leftIcon={<HelpIcon />} onClick={handleShowInstructions}>
+              Instructions
+            </Button>
+            <Button colorScheme={"red"} leftIcon={<XIcon />} onClick={handleClearVideos}>
+              Remove all videos
+            </Button>
+          </ButtonGroup>
+        </Flex>
         <Grid templateColumns={`repeat(${rowCount}, 1fr)`} gap={0}>
           {renderedVideos}
           <GridItem>
@@ -108,6 +105,35 @@ export default function SetupVideos() {
           </GridItem>
         </Grid>
       </Box>
-    </WithSidebar>
+      <Modal isOpen={showSetupInstructions} onClose={handleCloseInstructions}>
+        <ModalOverlay />
+        <ModalContent>
+          <ModalHeader>How to use</ModalHeader>
+          <ModalCloseButton />
+          <ModalBody>
+            <Heading color={"white"} fontSize={"lg"} mb={"4"}>
+              Getting started
+            </Heading>
+            <OrderedList my={"4"}>
+              <ListItem mb={"4"}>Either click to add videos or drag and drop them from your desktop.</ListItem>
+              <ListItem mb={"4"}>
+                Pick a point in time that is easy to identify in all the videos (countdowns before a round start works
+                well here)
+              </ListItem>
+              <ListItem mb={"4"}>
+                Align all the videos to the same point in time using the controls on each video.
+              </ListItem>
+              <ListItem mb={"4"}>When all videos are aligned, go to the "Review" tab to start the VOD review</ListItem>
+            </OrderedList>
+          </ModalBody>
+
+          <ModalFooter>
+            <Button colorScheme="blue" onClick={handleCloseInstructions}>
+              Close
+            </Button>
+          </ModalFooter>
+        </ModalContent>
+      </Modal>
+    </>
   );
 }